# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chpok API - civic incident reporting platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports:
    get:
      operationId: getReports
      tags: [reports]
      summary: Get reports (user's own, or all for admin)
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ReportStatus"
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/ProviderCategoryEnum"
        - name: providerId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportListResponse"

    post:
      operationId: createReport
      tags: [reports]
      summary: Submit a new incident report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportInput"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a single report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateReport
      tags: [reports]
      summary: Update an existing incident report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportInput"
      responses:
        "200":
          description: Report updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── MEDIA UPLOADS ───────────────────────────────────────────────────────────
  /reports/{reportId}/media:
    post:
      operationId: addReportMedia
      tags: [media]
      summary: Attach media metadata to a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMediaInput"
      responses:
        "201":
          description: Media record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportMedia"

  # ─── ADMIN ROUTES ─────────────────────────────────────────────────────────────
  /admin/reports:
    get:
      operationId: adminGetReports
      tags: [admin]
      summary: Admin - get all reports with full filters
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ReportStatus"
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/ProviderCategoryEnum"
        - name: providerId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, status, category]
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated admin report list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminReportListResponse"
  /admin/reports/{id}:
    get:
      operationId: adminGetReport
      tags: [admin]
      summary: Admin - get full report detail with history and notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full admin report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminReportDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: adminUpdateReportStatus
      tags: [admin]
      summary: Admin - update report status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateStatusInput"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /admin/reports/{id}/notes:
    post:
      operationId: adminAddNote
      tags: [admin]
      summary: Admin - add a note to a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddNoteInput"
      responses:
        "201":
          description: Note added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminNote"

  # ─── STATS ───────────────────────────────────────────────────────────────────
...
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get current user profile
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

    post:
      operationId: upsertUserProfile
      tags: [users]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    # ─── ENUMS ────────────────────────────────────────────────────────────────
    ProviderCategoryEnum:
      type: string
      enum:
        - delivery
        - micromobility
        - carsharing
        - taxi
        - car
        - other
    ReportStatus:
      type: string
      enum:
        - new
        - in_review
        - confirmed
        - rejected
        - resolved
        - archived
...
    Report:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
          nullable: true
        isAnonymous:
          type: boolean
        category:
          $ref: "#/components/schemas/ProviderCategoryEnum"
        providerId:
          type: string
        providerLabel:
          type: string
        description:
          type: string
        voiceNotePath:
          type: string
          nullable: true
        transcriptRaw:
          type: string
          nullable: true
        transcriptClean:
          type: string
          nullable: true
        transcriptStatus:
          type: string
        transcriptLanguage:
          type: string
          nullable: true
        transcriptProvider:
          type: string
          nullable: true
        transcriptError:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ReportStatus"
        deviceGeo:
          $ref: "#/components/schemas/GeoPoint"
          nullable: true
        addressText:
          type: string
          nullable: true
        mediaCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - isAnonymous
        - category
        - providerId
        - providerLabel
        - description
        - transcriptStatus
        - status
        - mediaCount
        - createdAt
        - updatedAt
    ReportDetail:
      allOf:
        - $ref: "#/components/schemas/Report"
        - type: object
          properties:
            media:
              type: array
              items:
                $ref: "#/components/schemas/ReportMedia"
            deviceContext:
              type: object
              nullable: true
              properties:
                platform:
                  type: string
                  nullable: true
                osVersion:
                  type: string
                  nullable: true
                appVersion:
                  type: string
                  nullable: true
          required: [media]
    ReportListResponse:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/Report"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasMore:
          type: boolean
      required: [reports, total, page, limit, hasMore]
    CreateReportInput:
      type: object
      properties:
        userId:
          type: string
          nullable: true
        isAnonymous:
          type: boolean
        category:
          $ref: "#/components/schemas/ProviderCategoryEnum"
        providerId:
          type: string
        description:
          type: string
        voiceNotePath:
          type: string
          nullable: true
        transcriptRaw:
          type: string
          nullable: true
        transcriptClean:
          type: string
          nullable: true
        transcriptStatus:
          type: string
        transcriptLanguage:
          type: string
          nullable: true
        transcriptProvider:
          type: string
          nullable: true
        transcriptError:
          type: string
          nullable: true
        deviceGeo:
          $ref: "#/components/schemas/GeoPoint"
          nullable: true
        addressText:
          type: string
          nullable: true
        deviceContext:
          type: object
          nullable: true
          properties:
            platform:
              type: string
              nullable: true
            osVersion:
              type: string
              nullable: true
            appVersion:
              type: string
              nullable: true
      required: [isAnonymous, category, providerId, description]
...
    AdminUpdateStatusInput:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ReportStatus"
        note:
          type: string
          nullable: true
        adminId:
          type: string
          nullable: true
      required: [status]
...